requirements.txt
Django>=5.1
psycopg2-binary>=2.9
django-environ>=0.11
openpyxl>=3.1
reportlab>=4.0
weasyprint>=60
qrcode[pil]>=7.4
Pillow>=10.1
requests>=2.31
django-ratelimit>=4.1
gunicorn>=21.2
whitenoise>=6.6
